Delete IntroR2021_Evaluation.pdf
I love this document. Will take it down so folks don't have it in their downloads for Intro to R in Nov. I've put it up one OneDrive in the Feb2022 Into to R folder as a backup. All this to say, I think we should consider how we can use this platform as a repository for SME materials or other such course content also.